+++ server/routes/smell/previous.js
@@ -1,18 +1,8 @@
 import constants from '../../utils/constants.js'
 
 const handlers = {
-  get: (_request, h) => {
-    const context = _getContext()
-    return h.view(constants.views.SMELL_PREVIOUS, {
-      ...context
-    })
-  }
-}
-
-const _getContext = () => {
-  return {
-    pageTitle: 'Report an environmental incident',
-    hideBackLink: false
+  get: (_request, _h) => {
+    return 'hello world'
   }
 }
