+++ server/views/smell/location-description.html
